# Optimize Array#many? without a block